implemented trascription and extraction
