--- testing_individual_parts/Archive/conversation.py
@@ -1,6 +1,0 @@
-CONVERSATION = '''Hello my name is Jason Jones, how may I assist you. 
-May I please have your first and last name. Hello I'm Donald Trump. I'm looking to reset my password. Ok Donald, I'll be happy to help you. 
-Could you please provide me with your identification date of birth. Sure, my date of birth is July 15th 1946. 
-Awesome I need one more piece of information to verify something from you Donald. What is your mother's maiden name? 
-Sure my mother's maiden name was MacLeod. Ok cool, I could verify you Mr Trump. Your new password is elephant129. 
-You should be all set to go. Thanks alot! Ok Thank you!'''
